MAP-6 Enhanced login functionality and improved error handling
In this commit, we have made several updates to improve the login functionality and error handling. The `Main` function in `App.js` now includes a title option for the `Login` screen. The `loginUser` function in `accountApi.js` has been modified to delete any existing JWT token before attempting to log in and now returns a boolean value indicating the success or failure of the login attempt.

The `postData` function in `apiService.js` and several functions in `jwtService.js` have been updated to remove console log statements. A new function `jwtKillToken` has been added to `jwtService.js` to delete the JWT token from secure storage.

Significant updates have been made to the `LoginScreen` component in `loginScreen.js`. It now includes state variables for login failure and form validation errors. The `handleLoginPress` function validates the form before attempting to log in and navigates to the `Home` screen upon successful login. A new `validateForm` function checks that the username and password fields are not empty. The `Forgot Password` button has been changed from a `TouchableOpacity` to a `Button`.