MAP-9
Title: Added CreateAccountScreen and updated related components

Summary:
A new `CreateAccountScreen` component was added to the navigation stack in `App.js`, allowing users to navigate from the login screen to the account creation screen. Two new functions, `isDisplayNameAvailable` and `isEmailAvailable`, were added to `accountApi.js` to check the availability of display names and emails for new users. A `debounce` function was added to `apiService.js` to prevent excessive requests. The `CreateAccountScreen` component was implemented in `createAccountScreen.js`, including form fields and validation for new account creation. The `handleCreateAccountPress` function in `loginScreen.js` was updated to navigate to the `CreateAccountScreen`, and the login button was changed to a `Pressable` component for better customization. Lastly, the `separator`, `button`, and `buttonText` styles in `globalStyles.js` were updated to match the new design.